gitindex: move package from internal
