+++ cmd/zoekt-git-index/main.go
@@ -28,8 +28,8 @@ import (
 	"go.uber.org/automaxprocs/maxprocs"
 
 	"github.com/sourcegraph/zoekt/cmd"
+	"github.com/sourcegraph/zoekt/gitindex"
 	"github.com/sourcegraph/zoekt/internal/ctags"
-	"github.com/sourcegraph/zoekt/internal/gitindex"
 	"github.com/sourcegraph/zoekt/internal/profiler"
 )
