+++ cmd/zoekt-mirror-gitea/main.go
@@ -30,7 +30,7 @@ import (
 
 	"code.gitea.io/sdk/gitea"
 
-	"github.com/sourcegraph/zoekt/internal/gitindex"
+	"github.com/sourcegraph/zoekt/gitindex"
 )
 
 type topicsFlag []string
