+++ cmd/zoekt-mirror-github/main.go
@@ -33,7 +33,7 @@ import (
 	"github.com/google/go-github/v27/github"
 	"golang.org/x/oauth2"
 
-	"github.com/sourcegraph/zoekt/internal/gitindex"
+	"github.com/sourcegraph/zoekt/gitindex"
 )
 
 type topicsFlag []string
