+++ gitindex/clone.go
