+++ gitindex/clone_test.go
