+++ gitindex/delete.go
