+++ gitindex/delete_test.go
