+++ gitindex/filter.go
