+++ gitindex/repocache_test.go
